Update 5 - Div
Wasn't sure what the purpose of div was, so I looked at my website and realized that the input fields were adjacent to eachother, therefore looking scrambled and unorganized.

Div helps create the "division" needed for text field to remain separate and organized.

Updated the syntax, and now the form is looking correct.