Update 6.01
Checked for differences between the final website and my current website, noticed that the input fields had a space in the final version. Returned to my index and added this space that was missing between the label and the input tags.

Also noticed that I was missing the Age header for the dropdown, added it before the options. The proper title is there now.